Tag the 1.6.3 release

+++ src/crashrpt/excprpt.cpp
@@ -222,7 +222,7 @@ CleanUp:
 //
 int CExceptionReport::getNumSymbolFiles()
 {
-   return (int)m_symFiles.size();
+   return m_symFiles.size();
 }
